moving api errors to common lib
