+++ libs/api/src/errors/assertions/index.ts
@@ -1,3 +1,1 @@
-export { default as isApplicationServerError } from './isApplicationServerError';
-export { default as isApplicationServerErrorLike } from './isApplicationServerErrorLike';
-export { default as isHttpErrorLike } from './isHttpErrorLike';
+export * from '@node/errors/assertions';
