+++ libs/api/src/errors/handlers/index.ts
@@ -1,6 +1,1 @@
-export { default as makeThirdPartyErrorHandler } from './makeThirdPartyErrorHandler';
-export { default as makeInternalErrorHandler } from './makeInternalErrorHandler';
-export { default as makeUnauthorizedErrorHandler } from './makeUnauthorizedErrorHandler';
-export { default as makeBadRequestErrorHandler } from './makeBadRequestErrorHandler';
-export { default as makeNotFoundErrorHandler } from './makeNotFoundErrorHandler';
-export { default as makeVideoApiErrorHandler } from './makeVideoApiErrorHandler';
+export * from '@node/errors/handlers';
