+++ libs/api/src/errors/helpers/index.ts
@@ -1,2 +1,2 @@
-export { default as mapServerSourceToState } from './mapServerSourceToState';
+export * from '@node/errors/helpers';
 export { default as toTRPCError } from './toTRPCError';
