+++ libs/api/src/executions/index.ts
@@ -1,1 +1,1 @@
-export { default as assertResult } from './assertResult';
+export * from '@node/executions';
