+++ libs/common/srcNode/errors/handlers/makeNotFoundErrorHandler/makeNotFoundErrorHandler.ts
