+++ libs/common/srcNode/errors/handlers/makeThirdPartyErrorHandler/makeThirdPartyErrorHandler.ts
