+++ libs/common/srcNode/errors/handlers/makeVideoApiErrorHandler/makeVideoApiErrorHandler.ts
