+++ libs/common/srcNode/errors/helpers/mapServerSourceToState/mapServerSourceToState.ts
