+++ libs/common/srcNode/errors/index.ts
@@ -1,4 +1,6 @@
-export * from '@node/errors';
+export * from '../../src/errors';
+
+export * from './ApplicationServerError';
 export * from './assertions';
 export * from './handlers';
 export * from './helpers';
