+++ libs/common/srcNode/errors/types/ApplicationErrorFallbackConfig.ts
