+++ libs/common/srcNode/errors/types/ApplicationErrorState.ts
