+++ libs/common/srcNode/errors/types/ErrorSeverity.ts
