+++ libs/common/srcNode/errors/types/HttpError.ts
