+++ libs/common/srcNode/executions/index.ts
@@ -1,0 +1,4 @@
+export * from '../../src/execution';
+
+// Server-specific assertResult shadows the base one
+export { default as assertResult } from './assertResult';
