+++ libs/common/srcNode/index.ts
@@ -1,7 +1,12 @@
 export * from '../src';
 
 export * from './assertions';
+export * from './errors';
+export * from './executions';
 export * from './helpers';
 export * from './routing';
 export * from './schemas';
 export * from './types';
+
+// Server-specific assertResult shadows the base one from @common/execution
+export { assertResult } from './executions';
